\name{birds}
\alias{birds}
\docType{data}
\title{
Breeding Bird Survey
}
\description{
This dataset is a subset of the Breeding Bird Survey.
}
\usage{data("birds")}
\format{
  A data frame with 13608 observations and 384 variables.
  \describe{
\item{\code{loc.id}}{Location index}
    \item{\code{aou530}}{Bird specie 530 presence (1) or absence (0)}
    \item{\code{aou540}}{Bird specie 540 presence (1) or absence (0)}
    \item{\code{aou590}}{Bird specie 590 presence (1) or absence (0)}
    \item{\code{aou1200}}{Bird specie 1200 presence (1) or absence (0)}
    \item{\code{aou1320}}{Bird specie 1320 presence (1) or absence (0)}
    \item{\code{aou1400}}{Bird specie 1400 presence (1) or absence (0)}
    \item{\code{aou1420}}{Bird specie 1420 presence (1) or absence (0)}
    \item{\code{aou2140}}{Bird specie 2140 presence (1) or absence (0)}
    \item{\code{aou2300}}{Bird specie 2300 presence (1) or absence (0)}
    \item{\code{aou2490}}{Bird specie 2490 presence (1) or absence (0)}
    \item{\code{aou2580}}{Bird specie 2580 presence (1) or absence (0)}
    \item{\code{aou2640}}{Bird specie 2640 presence (1) or absence (0)}
    \item{\code{aou2730}}{Bird specie 2730 presence (1) or absence (0)}
    \item{\code{aou3091}}{Bird specie 3091 presence (1) or absence (0)}
    \item{\code{aou3160}}{Bird specie 3160 presence (1) or absence (0)}
    \item{\code{aou3310}}{Bird specie 3310 presence (1) or absence (0)}
    \item{\code{aou3370}}{Bird specie 3370 presence (1) or absence (0)}
    \item{\code{aou4440}}{Bird specie 4440 presence (1) or absence (0)}
    \item{\code{aou4740}}{Bird specie 4740 presence (1) or absence (0)}
    \item{\code{aou4750}}{Bird specie 4750 presence (1) or absence (0)}
    \item{\code{aou4880}}{Bird specie 4880 presence (1) or absence (0)}
    \item{\code{aou4930}}{Bird specie 4930 presence (1) or absence (0)}
    \item{\code{aou4950}}{Bird specie 4950 presence (1) or absence (0)}
    \item{\code{aou4970}}{Bird specie 4970 presence (1) or absence (0)}
    \item{\code{aou4980}}{Bird specie 4980 presence (1) or absence (0)}
    \item{\code{aou5011}}{Bird specie 5011 presence (1) or absence (0)}
    \item{\code{aou5100}}{Bird specie 5100 presence (1) or absence (0)}
    \item{\code{aou5290}}{Bird specie 5290 presence (1) or absence (0)}
    \item{\code{aou5380}}{Bird specie 5380 presence (1) or absence (0)}
    \item{\code{aou5400}}{Bird specie 5400 presence (1) or absence (0)}
    \item{\code{aou5420}}{Bird specie 5420 presence (1) or absence (0)}
    \item{\code{aou5610}}{Bird specie 5610 presence (1) or absence (0)}
    \item{\code{aou5620}}{Bird specie 5620 presence (1) or absence (0)}
    \item{\code{aou5810}}{Bird specie 5810 presence (1) or absence (0)}
    \item{\code{aou5880}}{Bird specie 5880 presence (1) or absence (0)}
    \item{\code{aou6050}}{Bird specie 6050 presence (1) or absence (0)}
    \item{\code{aou6130}}{Bird specie 6130 presence (1) or absence (0)}
    \item{\code{aou6270}}{Bird specie 6270 presence (1) or absence (0)}
    \item{\code{aou6520}}{Bird specie 6520 presence (1) or absence (0)}
    \item{\code{aou6810}}{Bird specie 6810 presence (1) or absence (0)}
    \item{\code{aou6882}}{Bird specie 6882 presence (1) or absence (0)}
    \item{\code{aou7610}}{Bird specie 7610 presence (1) or absence (0)}
    \item{\code{aou510}}{Bird specie 510 presence (1) or absence (0)}
    \item{\code{aou2210}}{Bird specie 2210 presence (1) or absence (0)}
    \item{\code{aou2250}}{Bird specie 2250 presence (1) or absence (0)}
    \item{\code{aou2610}}{Bird specie 2610 presence (1) or absence (0)}
    \item{\code{aou3131}}{Bird specie 3131 presence (1) or absence (0)}
    \item{\code{aou4470}}{Bird specie 4470 presence (1) or absence (0)}
    \item{\code{aou4570}}{Bird specie 4570 presence (1) or absence (0)}
    \item{\code{aou6120}}{Bird specie 6120 presence (1) or absence (0)}
    \item{\code{aou6140}}{Bird specie 6140 presence (1) or absence (0)}
    \item{\code{aou7250}}{Bird specie 7250 presence (1) or absence (0)}
    \item{\code{aou1350}}{Bird specie 1350 presence (1) or absence (0)}
    \item{\code{aou1410}}{Bird specie 1410 presence (1) or absence (0)}
    \item{\code{aou1720}}{Bird specie 1720 presence (1) or absence (0)}
    \item{\code{aou2630}}{Bird specie 2630 presence (1) or absence (0)}
    \item{\code{aou3080}}{Bird specie 3080 presence (1) or absence (0)}
    \item{\code{aou3420}}{Bird specie 3420 presence (1) or absence (0)}
    \item{\code{aou4123}}{Bird specie 4123 presence (1) or absence (0)}
    \item{\code{aou4670}}{Bird specie 4670 presence (1) or absence (0)}
    \item{\code{aou5110}}{Bird specie 5110 presence (1) or absence (0)}
    \item{\code{aou5460}}{Bird specie 5460 presence (1) or absence (0)}
    \item{\code{aou7040}}{Bird specie 7040 presence (1) or absence (0)}
    \item{\code{aou7210}}{Bird specie 7210 presence (1) or absence (0)}
    \item{\code{aou22860}}{Bird specie 22860 presence (1) or absence (0)}
    \item{\code{aou40}}{Bird specie 40 presence (1) or absence (0)}
    \item{\code{aou60}}{Bird specie 60 presence (1) or absence (0)}
    \item{\code{aou1250}}{Bird specie 1250 presence (1) or absence (0)}
    \item{\code{aou1430}}{Bird specie 1430 presence (1) or absence (0)}
    \item{\code{aou1490}}{Bird specie 1490 presence (1) or absence (0)}
    \item{\code{aou2240}}{Bird specie 2240 presence (1) or absence (0)}
    \item{\code{aou2881}}{Bird specie 2881 presence (1) or absence (0)}
    \item{\code{aou3750}}{Bird specie 3750 presence (1) or absence (0)}
    \item{\code{aou4620}}{Bird specie 4620 presence (1) or absence (0)}
    \item{\code{aou1370}}{Bird specie 1370 presence (1) or absence (0)}
    \item{\code{aou4200}}{Bird specie 4200 presence (1) or absence (0)}
    \item{\code{aou5520}}{Bird specie 5520 presence (1) or absence (0)}
    \item{\code{aou6190}}{Bird specie 6190 presence (1) or absence (0)}
    \item{\code{aou7050}}{Bird specie 7050 presence (1) or absence (0)}
    \item{\code{aou7150}}{Bird specie 7150 presence (1) or absence (0)}
    \item{\code{aou1670}}{Bird specie 1670 presence (1) or absence (0)}
    \item{\code{aou1940}}{Bird specie 1940 presence (1) or absence (0)}
    \item{\code{aou4120}}{Bird specie 4120 presence (1) or absence (0)}
    \item{\code{aou5070}}{Bird specie 5070 presence (1) or absence (0)}
    \item{\code{aou3480}}{Bird specie 3480 presence (1) or absence (0)}
    \item{\code{aou5190}}{Bird specie 5190 presence (1) or absence (0)}
    \item{\code{aou5990}}{Bird specie 5990 presence (1) or absence (0)}
    \item{\code{aou10}}{Bird specie 10 presence (1) or absence (0)}
    \item{\code{aou6220}}{Bird specie 6220 presence (1) or absence (0)}
    \item{\code{aou70}}{Bird specie 70 presence (1) or absence (0)}
    \item{\code{aou4860}}{Bird specie 4860 presence (1) or absence (0)}
    \item{\code{aou6160}}{Bird specie 6160 presence (1) or absence (0)}
    \item{\code{aou1470}}{Bird specie 1470 presence (1) or absence (0)}
    \item{\code{aou2260}}{Bird specie 2260 presence (1) or absence (0)}
    \item{\code{aou640}}{Bird specie 640 presence (1) or absence (0)}
    \item{\code{aou5960}}{Bird specie 5960 presence (1) or absence (0)}
    \item{\code{aou1390}}{Bird specie 1390 presence (1) or absence (0)}
    \item{\code{aou1460}}{Bird specie 1460 presence (1) or absence (0)}
    \item{\code{aou3600}}{Bird specie 3600 presence (1) or absence (0)}
    \item{\code{aou6460}}{Bird specie 6460 presence (1) or absence (0)}
    \item{\code{aou5600}}{Bird specie 5600 presence (1) or absence (0)}
    \item{\code{aou7280}}{Bird specie 7280 presence (1) or absence (0)}
    \item{\code{aou7350}}{Bird specie 7350 presence (1) or absence (0)}
    \item{\code{aou3640}}{Bird specie 3640 presence (1) or absence (0)}
    \item{\code{aou3490}}{Bird specie 3490 presence (1) or absence (0)}
    \item{\code{aou3940}}{Bird specie 3940 presence (1) or absence (0)}
    \item{\code{aou3550}}{Bird specie 3550 presence (1) or absence (0)}
    \item{\code{aou3670}}{Bird specie 3670 presence (1) or absence (0)}
    \item{\code{aou7580}}{Bird specie 7580 presence (1) or absence (0)}
    \item{\code{aou6070}}{Bird specie 6070 presence (1) or absence (0)}
    \item{\code{aou7680}}{Bird specie 7680 presence (1) or absence (0)}
    \item{\code{aou4130}}{Bird specie 4130 presence (1) or absence (0)}
    \item{\code{aou3930}}{Bird specie 3930 presence (1) or absence (0)}
    \item{\code{aou5480}}{Bird specie 5480 presence (1) or absence (0)}
    \item{\code{aou3520}}{Bird specie 3520 presence (1) or absence (0)}
    \item{\code{aou1290}}{Bird specie 1290 presence (1) or absence (0)}
    \item{\code{aou6170}}{Bird specie 6170 presence (1) or absence (0)}
    \item{\code{aou3570}}{Bird specie 3570 presence (1) or absence (0)}
    \item{\code{aou690}}{Bird specie 690 presence (1) or absence (0)}
    \item{\code{aou2020}}{Bird specie 2020 presence (1) or absence (0)}
    \item{\code{aou770}}{Bird specie 770 presence (1) or absence (0)}
    \item{\code{aou3000}}{Bird specie 3000 presence (1) or absence (0)}
    \item{\code{aou3900}}{Bird specie 3900 presence (1) or absence (0)}
    \item{\code{aou4641}}{Bird specie 4641 presence (1) or absence (0)}
    \item{\code{aou4661}}{Bird specie 4661 presence (1) or absence (0)}
    \item{\code{aou4690}}{Bird specie 4690 presence (1) or absence (0)}
    \item{\code{aou4840}}{Bird specie 4840 presence (1) or absence (0)}
    \item{\code{aou5330}}{Bird specie 5330 presence (1) or absence (0)}
    \item{\code{aou5540}}{Bird specie 5540 presence (1) or absence (0)}
    \item{\code{aou5670}}{Bird specie 5670 presence (1) or absence (0)}
    \item{\code{aou5830}}{Bird specie 5830 presence (1) or absence (0)}
    \item{\code{aou5850}}{Bird specie 5850 presence (1) or absence (0)}
    \item{\code{aou6560}}{Bird specie 6560 presence (1) or absence (0)}
    \item{\code{aou6750}}{Bird specie 6750 presence (1) or absence (0)}
    \item{\code{aou6800}}{Bird specie 6800 presence (1) or absence (0)}
    \item{\code{aou6850}}{Bird specie 6850 presence (1) or absence (0)}
    \item{\code{aou7490}}{Bird specie 7490 presence (1) or absence (0)}
    \item{\code{aou7630}}{Bird specie 7630 presence (1) or absence (0)}
    \item{\code{aou4021}}{Bird specie 4021 presence (1) or absence (0)}
    \item{\code{aou4590}}{Bird specie 4590 presence (1) or absence (0)}
    \item{\code{aou4660}}{Bird specie 4660 presence (1) or absence (0)}
    \item{\code{aou7380}}{Bird specie 7380 presence (1) or absence (0)}
    \item{\code{aou7670}}{Bird specie 7670 presence (1) or absence (0)}
    \item{\code{aou4910}}{Bird specie 4910 presence (1) or absence (0)}
    \item{\code{aou7540}}{Bird specie 7540 presence (1) or absence (0)}
    \item{\code{aou4330}}{Bird specie 4330 presence (1) or absence (0)}
    \item{\code{aou6240}}{Bird specie 6240 presence (1) or absence (0)}
    \item{\code{aou7480}}{Bird specie 7480 presence (1) or absence (0)}
    \item{\code{aou7400}}{Bird specie 7400 presence (1) or absence (0)}
    \item{\code{aou5210}}{Bird specie 5210 presence (1) or absence (0)}
    \item{\code{aou6680}}{Bird specie 6680 presence (1) or absence (0)}
    \item{\code{aou6740}}{Bird specie 6740 presence (1) or absence (0)}
    \item{\code{aou5671}}{Bird specie 5671 presence (1) or absence (0)}
    \item{\code{aou6291}}{Bird specie 6291 presence (1) or absence (0)}
    \item{\code{aou6870}}{Bird specie 6870 presence (1) or absence (0)}
    \item{\code{aou4940}}{Bird specie 4940 presence (1) or absence (0)}
    \item{\code{aou2120}}{Bird specie 2120 presence (1) or absence (0)}
    \item{\code{aou7030}}{Bird specie 7030 presence (1) or absence (0)}
    \item{\code{aou2060}}{Bird specie 2060 presence (1) or absence (0)}
    \item{\code{aou3330}}{Bird specie 3330 presence (1) or absence (0)}
    \item{\code{aou1900}}{Bird specie 1900 presence (1) or absence (0)}
    \item{\code{aou3250}}{Bird specie 3250 presence (1) or absence (0)}
    \item{\code{aou7590}}{Bird specie 7590 presence (1) or absence (0)}
    \item{\code{aou4020}}{Bird specie 4020 presence (1) or absence (0)}
    \item{\code{aou5580}}{Bird specie 5580 presence (1) or absence (0)}
    \item{\code{aou1440}}{Bird specie 1440 presence (1) or absence (0)}
    \item{\code{aou4560}}{Bird specie 4560 presence (1) or absence (0)}
    \item{\code{aou4770}}{Bird specie 4770 presence (1) or absence (0)}
    \item{\code{aou6550}}{Bird specie 6550 presence (1) or absence (0)}
    \item{\code{aou5220}}{Bird specie 5220 presence (1) or absence (0)}
    \item{\code{aou5950}}{Bird specie 5950 presence (1) or absence (0)}
    \item{\code{aou5140}}{Bird specie 5140 presence (1) or absence (0)}
    \item{\code{aou7560}}{Bird specie 7560 presence (1) or absence (0)}
    \item{\code{aou6290}}{Bird specie 6290 presence (1) or absence (0)}
    \item{\code{aou5170}}{Bird specie 5170 presence (1) or absence (0)}
    \item{\code{aou7270}}{Bird specie 7270 presence (1) or absence (0)}
    \item{\code{aou1310}}{Bird specie 1310 presence (1) or absence (0)}
    \item{\code{aou3920}}{Bird specie 3920 presence (1) or absence (0)}
    \item{\code{aou4050}}{Bird specie 4050 presence (1) or absence (0)}
    \item{\code{aou4680}}{Bird specie 4680 presence (1) or absence (0)}
    \item{\code{aou6150}}{Bird specie 6150 presence (1) or absence (0)}
    \item{\code{aou4010}}{Bird specie 4010 presence (1) or absence (0)}
    \item{\code{aou7221}}{Bird specie 7221 presence (1) or absence (0)}
    \item{\code{aou6610}}{Bird specie 6610 presence (1) or absence (0)}
    \item{\code{aou5150}}{Bird specie 5150 presence (1) or absence (0)}
    \item{\code{aou7260}}{Bird specie 7260 presence (1) or absence (0)}
    \item{\code{aou3880}}{Bird specie 3880 presence (1) or absence (0)}
    \item{\code{aou3320}}{Bird specie 3320 presence (1) or absence (0)}
    \item{\code{aou1500}}{Bird specie 1500 presence (1) or absence (0)}
    \item{\code{aou4520}}{Bird specie 4520 presence (1) or absence (0)}
    \item{\code{aou6790}}{Bird specie 6790 presence (1) or absence (0)}
    \item{\code{aou6110}}{Bird specie 6110 presence (1) or absence (0)}
    \item{\code{aou4280}}{Bird specie 4280 presence (1) or absence (0)}
    \item{\code{aou5840}}{Bird specie 5840 presence (1) or absence (0)}
    \item{\code{aou1810}}{Bird specie 1810 presence (1) or absence (0)}
    \item{\code{aou6570}}{Bird specie 6570 presence (1) or absence (0)}
    \item{\code{aou7010}}{Bird specie 7010 presence (1) or absence (0)}
    \item{\code{aou3430}}{Bird specie 3430 presence (1) or absence (0)}
    \item{\code{aou6360}}{Bird specie 6360 presence (1) or absence (0)}
    \item{\code{aou4000}}{Bird specie 4000 presence (1) or absence (0)}
    \item{\code{aou6860}}{Bird specie 6860 presence (1) or absence (0)}
    \item{\code{aou7222}}{Bird specie 7222 presence (1) or absence (0)}
    \item{\code{aou6670}}{Bird specie 6670 presence (1) or absence (0)}
    \item{\code{aou4630}}{Bird specie 4630 presence (1) or absence (0)}
    \item{\code{aou3680}}{Bird specie 3680 presence (1) or absence (0)}
    \item{\code{aou550}}{Bird specie 550 presence (1) or absence (0)}
    \item{\code{aou4402}}{Bird specie 4402 presence (1) or absence (0)}
    \item{\code{aou6450}}{Bird specie 6450 presence (1) or absence (0)}
    \item{\code{aou6830}}{Bird specie 6830 presence (1) or absence (0)}
    \item{\code{aou4780}}{Bird specie 4780 presence (1) or absence (0)}
    \item{\code{aou3780}}{Bird specie 3780 presence (1) or absence (0)}
    \item{\code{aou7240}}{Bird specie 7240 presence (1) or absence (0)}
    \item{\code{aou1870}}{Bird specie 1870 presence (1) or absence (0)}
    \item{\code{aou5080}}{Bird specie 5080 presence (1) or absence (0)}
    \item{\code{aou5012}}{Bird specie 5012 presence (1) or absence (0)}
    \item{\code{aou4060}}{Bird specie 4060 presence (1) or absence (0)}
    \item{\code{aou4640}}{Bird specie 4640 presence (1) or absence (0)}
    \item{\code{aou5590}}{Bird specie 5590 presence (1) or absence (0)}
    \item{\code{aou6590}}{Bird specie 6590 presence (1) or absence (0)}
    \item{\code{aou3100}}{Bird specie 3100 presence (1) or absence (0)}
    \item{\code{aou5180}}{Bird specie 5180 presence (1) or absence (0)}
    \item{\code{aou7570}}{Bird specie 7570 presence (1) or absence (0)}
    \item{\code{aou7660}}{Bird specie 7660 presence (1) or absence (0)}
    \item{\code{aou2010}}{Bird specie 2010 presence (1) or absence (0)}
    \item{\code{aou2890}}{Bird specie 2890 presence (1) or absence (0)}
    \item{\code{aou3870}}{Bird specie 3870 presence (1) or absence (0)}
    \item{\code{aou4090}}{Bird specie 4090 presence (1) or absence (0)}
    \item{\code{aou4160}}{Bird specie 4160 presence (1) or absence (0)}
    \item{\code{aou4230}}{Bird specie 4230 presence (1) or absence (0)}
    \item{\code{aou4610}}{Bird specie 4610 presence (1) or absence (0)}
    \item{\code{aou4650}}{Bird specie 4650 presence (1) or absence (0)}
    \item{\code{aou5010}}{Bird specie 5010 presence (1) or absence (0)}
    \item{\code{aou5630}}{Bird specie 5630 presence (1) or absence (0)}
    \item{\code{aou5870}}{Bird specie 5870 presence (1) or absence (0)}
    \item{\code{aou5930}}{Bird specie 5930 presence (1) or absence (0)}
    \item{\code{aou5970}}{Bird specie 5970 presence (1) or absence (0)}
    \item{\code{aou5980}}{Bird specie 5980 presence (1) or absence (0)}
    \item{\code{aou6040}}{Bird specie 6040 presence (1) or absence (0)}
    \item{\code{aou6080}}{Bird specie 6080 presence (1) or absence (0)}
    \item{\code{aou6100}}{Bird specie 6100 presence (1) or absence (0)}
    \item{\code{aou6280}}{Bird specie 6280 presence (1) or absence (0)}
    \item{\code{aou6310}}{Bird specie 6310 presence (1) or absence (0)}
    \item{\code{aou6710}}{Bird specie 6710 presence (1) or absence (0)}
    \item{\code{aou6730}}{Bird specie 6730 presence (1) or absence (0)}
    \item{\code{aou7180}}{Bird specie 7180 presence (1) or absence (0)}
    \item{\code{aou7310}}{Bird specie 7310 presence (1) or absence (0)}
    \item{\code{aou7360}}{Bird specie 7360 presence (1) or absence (0)}
    \item{\code{aou7510}}{Bird specie 7510 presence (1) or absence (0)}
    \item{\code{aou7550}}{Bird specie 7550 presence (1) or absence (0)}
    \item{\code{aou5060}}{Bird specie 5060 presence (1) or absence (0)}
    \item{\code{aou6480}}{Bird specie 6480 presence (1) or absence (0)}
    \item{\code{aou6770}}{Bird specie 6770 presence (1) or absence (0)}
    \item{\code{aou6840}}{Bird specie 6840 presence (1) or absence (0)}
    \item{\code{aou3390}}{Bird specie 3390 presence (1) or absence (0)}
    \item{\code{aou6370}}{Bird specie 6370 presence (1) or absence (0)}
    \item{\code{aou6630}}{Bird specie 6630 presence (1) or absence (0)}
    \item{\code{aou3260}}{Bird specie 3260 presence (1) or absence (0)}
    \item{\code{aou6380}}{Bird specie 6380 presence (1) or absence (0)}
    \item{\code{aou7290}}{Bird specie 7290 presence (1) or absence (0)}
    \item{\code{aou6390}}{Bird specie 6390 presence (1) or absence (0)}
    \item{\code{aou1960}}{Bird specie 1960 presence (1) or absence (0)}
    \item{\code{aou2030}}{Bird specie 2030 presence (1) or absence (0)}
    \item{\code{aou2000}}{Bird specie 2000 presence (1) or absence (0)}
    \item{\code{aou4900}}{Bird specie 4900 presence (1) or absence (0)}
    \item{\code{aou3730}}{Bird specie 3730 presence (1) or absence (0)}
    \item{\code{aou6760}}{Bird specie 6760 presence (1) or absence (0)}
    \item{\code{aou6410}}{Bird specie 6410 presence (1) or absence (0)}
    \item{\code{aou2001}}{Bird specie 2001 presence (1) or absence (0)}
    \item{\code{aou4171}}{Bird specie 4171 presence (1) or absence (0)}
    \item{\code{aou6580}}{Bird specie 6580 presence (1) or absence (0)}
    \item{\code{aou1970}}{Bird specie 1970 presence (1) or absence (0)}
    \item{\code{aou2190}}{Bird specie 2190 presence (1) or absence (0)}
    \item{\code{aou3290}}{Bird specie 3290 presence (1) or absence (0)}
    \item{\code{aou1840}}{Bird specie 1840 presence (1) or absence (0)}
    \item{\code{aou1180}}{Bird specie 1180 presence (1) or absence (0)}
    \item{\code{aou5750}}{Bird specie 5750 presence (1) or absence (0)}
    \item{\code{aou3200}}{Bird specie 3200 presence (1) or absence (0)}
    \item{\code{aou1880}}{Bird specie 1880 presence (1) or absence (0)}
    \item{\code{aou4881}}{Bird specie 4881 presence (1) or absence (0)}
    \item{\code{aou3270}}{Bird specie 3270 presence (1) or absence (0)}
    \item{\code{aou580}}{Bird specie 580 presence (1) or absence (0)}
    \item{\code{aou740}}{Bird specie 740 presence (1) or absence (0)}
    \item{\code{aou5130}}{Bird specie 5130 presence (1) or absence (0)}
    \item{\code{aou1340}}{Bird specie 1340 presence (1) or absence (0)}
    \item{\code{aou1990}}{Bird specie 1990 presence (1) or absence (0)}
    \item{\code{aou2110}}{Bird specie 2110 presence (1) or absence (0)}
    \item{\code{aou3190}}{Bird specie 3190 presence (1) or absence (0)}
    \item{\code{aou6010}}{Bird specie 6010 presence (1) or absence (0)}
    \item{\code{aou1770}}{Bird specie 1770 presence (1) or absence (0)}
    \item{\code{aou5280}}{Bird specie 5280 presence (1) or absence (0)}
    \item{\code{aou440}}{Bird specie 440 presence (1) or absence (0)}
    \item{\code{aou4030}}{Bird specie 4030 presence (1) or absence (0)}
    \item{\code{aou7410}}{Bird specie 7410 presence (1) or absence (0)}
    \item{\code{aou2971}}{Bird specie 2971 presence (1) or absence (0)}
    \item{\code{aou4240}}{Bird specie 4240 presence (1) or absence (0)}
    \item{\code{aou3120}}{Bird specie 3120 presence (1) or absence (0)}
    \item{\code{aou4679}}{Bird specie 4679 presence (1) or absence (0)}
    \item{\code{aou4080}}{Bird specie 4080 presence (1) or absence (0)}
    \item{\code{aou4540}}{Bird specie 4540 presence (1) or absence (0)}
    \item{\code{aou4691}}{Bird specie 4691 presence (1) or absence (0)}
    \item{\code{aou4920}}{Bird specie 4920 presence (1) or absence (0)}
    \item{\code{aou5040}}{Bird specie 5040 presence (1) or absence (0)}
    \item{\code{aou5730}}{Bird specie 5730 presence (1) or absence (0)}
    \item{\code{aou5738}}{Bird specie 5738 presence (1) or absence (0)}
    \item{\code{aou6292}}{Bird specie 6292 presence (1) or absence (0)}
    \item{\code{aou6340}}{Bird specie 6340 presence (1) or absence (0)}
    \item{\code{aou6650}}{Bird specie 6650 presence (1) or absence (0)}
    \item{\code{aou7190}}{Bird specie 7190 presence (1) or absence (0)}
    \item{\code{aou7331}}{Bird specie 7331 presence (1) or absence (0)}
    \item{\code{aou7430}}{Bird specie 7430 presence (1) or absence (0)}
    \item{\code{aou4320}}{Bird specie 4320 presence (1) or absence (0)}
    \item{\code{aou4480}}{Bird specie 4480 presence (1) or absence (0)}
    \item{\code{aou4810}}{Bird specie 4810 presence (1) or absence (0)}
    \item{\code{aou7020}}{Bird specie 7020 presence (1) or absence (0)}
    \item{\code{aou7070}}{Bird specie 7070 presence (1) or absence (0)}
    \item{\code{aou4180}}{Bird specie 4180 presence (1) or absence (0)}
    \item{\code{aou3960}}{Bird specie 3960 presence (1) or absence (0)}
    \item{\code{aou4250}}{Bird specie 4250 presence (1) or absence (0)}
    \item{\code{aou5900}}{Bird specie 5900 presence (1) or absence (0)}
    \item{\code{aou5650}}{Bird specie 5650 presence (1) or absence (0)}
    \item{\code{aou5300}}{Bird specie 5300 presence (1) or absence (0)}
    \item{\code{aou5690}}{Bird specie 5690 presence (1) or absence (0)}
    \item{\code{aou6090}}{Bird specie 6090 presence (1) or absence (0)}
    \item{\code{aou6640}}{Bird specie 6640 presence (1) or absence (0)}
    \item{\code{aou7300}}{Bird specie 7300 presence (1) or absence (0)}
    \item{\code{aou4070}}{Bird specie 4070 presence (1) or absence (0)}
    \item{\code{aou4040}}{Bird specie 4040 presence (1) or absence (0)}
    \item{\code{aou6440}}{Bird specie 6440 presence (1) or absence (0)}
    \item{\code{aou2950}}{Bird specie 2950 presence (1) or absence (0)}
    \item{\code{aou3850}}{Bird specie 3850 presence (1) or absence (0)}
    \item{\code{aou5120}}{Bird specie 5120 presence (1) or absence (0)}
    \item{\code{aou5800}}{Bird specie 5800 presence (1) or absence (0)}
    \item{\code{aou6200}}{Bird specie 6200 presence (1) or absence (0)}
    \item{\code{aou7130}}{Bird specie 7130 presence (1) or absence (0)}
    \item{\code{aou7460}}{Bird specie 7460 presence (1) or absence (0)}
    \item{\code{aou7520}}{Bird specie 7520 presence (1) or absence (0)}
    \item{\code{aou4210}}{Bird specie 4210 presence (1) or absence (0)}
    \item{\code{aou4290}}{Bird specie 4290 presence (1) or absence (0)}
    \item{\code{aou4530}}{Bird specie 4530 presence (1) or absence (0)}
    \item{\code{aou3210}}{Bird specie 3210 presence (1) or absence (0)}
    \item{\code{aou4310}}{Bird specie 4310 presence (1) or absence (0)}
    \item{\code{aou7170}}{Bird specie 7170 presence (1) or absence (0)}
    \item{\code{aou4110}}{Bird specie 4110 presence (1) or absence (0)}
    \item{\code{aou5910}}{Bird specie 5910 presence (1) or absence (0)}
    \item{\code{aou6430}}{Bird specie 6430 presence (1) or absence (0)}
    \item{\code{aou3350}}{Bird specie 3350 presence (1) or absence (0)}
    \item{\code{aou5050}}{Bird specie 5050 presence (1) or absence (0)}
    \item{\code{aou4710}}{Bird specie 4710 presence (1) or absence (0)}
    \item{\code{aou6320}}{Bird specie 6320 presence (1) or absence (0)}
    \item{\code{aou6330}}{Bird specie 6330 presence (1) or absence (0)}
    \item{\code{aou4960}}{Bird specie 4960 presence (1) or absence (0)}
    \item{\code{aou5940}}{Bird specie 5940 presence (1) or absence (0)}
    \item{\code{aou4580}}{Bird specie 4580 presence (1) or absence (0)}
    \item{\code{aou4870}}{Bird specie 4870 presence (1) or absence (0)}
    \item{\code{aou2930}}{Bird specie 2930 presence (1) or absence (0)}
    \item{\code{aou5780}}{Bird specie 5780 presence (1) or absence (0)}
    \item{\code{aou2882}}{Bird specie 2882 presence (1) or absence (0)}
    \item{\code{aou3620}}{Bird specie 3620 presence (1) or absence (0)}
    \item{\code{aou4430}}{Bird specie 4430 presence (1) or absence (0)}
    \item{\code{aou2940}}{Bird specie 2940 presence (1) or absence (0)}
    \item{\code{aou4360}}{Bird specie 4360 presence (1) or absence (0)}
    \item{\code{aou2920}}{Bird specie 2920 presence (1) or absence (0)}
    \item{\code{aou7420}}{Bird specie 7420 presence (1) or absence (0)}
    \item{\code{aou3970}}{Bird specie 3970 presence (1) or absence (0)}
    \item{\code{aou7330}}{Bird specie 7330 presence (1) or absence (0)}
    \item{\code{aou3990}}{Bird specie 3990 presence (1) or absence (0)}
    \item{\code{aou6690}}{Bird specie 6690 presence (1) or absence (0)}
    \item{\code{aou5911}}{Bird specie 5911 presence (1) or absence (0)}
    \item{\code{aou4760}}{Bird specie 4760 presence (1) or absence (0)}
    \item{\code{aou7100}}{Bird specie 7100 presence (1) or absence (0)}
    \item{\code{aou3090}}{Bird specie 3090 presence (1) or absence (0)}
    \item{\code{aou3050}}{Bird specie 3050 presence (1) or absence (0)}
    \item{\code{aou6620}}{Bird specie 6620 presence (1) or absence (0)}
    \item{\code{aou6540}}{Bird specie 6540 presence (1) or absence (0)}
    \item{\code{aou1830}}{Bird specie 1830 presence (1) or absence (0)}
    \item{\code{aou5470}}{Bird specie 5470 presence (1) or absence (0)}
    \item{\code{aou6883}}{Bird specie 6883 presence (1) or absence (0)}
    \item{\code{aou6420}}{Bird specie 6420 presence (1) or absence (0)}
    \item{\code{aou6121}}{Bird specie 6121 presence (1) or absence (0)}
    \item{\code{aou4100}}{Bird specie 4100 presence (1) or absence (0)}
    \item{\code{aou7320}}{Bird specie 7320 presence (1) or absence (0)}
    \item{\code{aou4461}}{Bird specie 4461 presence (1) or absence (0)}
    \item{\code{aou5860}}{Bird specie 5860 presence (1) or absence (0)}
  }
}
\source{Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding, K. Campbell and M.-A.R. Hudson. 2017. North American Breeding Bird Survey Dataset 1966 - 2016, version 2016.0. U.S. Geological Survey, Patuxent Wildlife Research Center. \url{https://www.pwrc.usgs.gov/bbs/RawData/}doi:10.5066/F7W0944J.}
\examples{
data(birds)
}
\keyword{datasets}
