\name{fishnet}
\alias{fishnet}
\docType{data}
\title{
Latitude and Longitude Fishnet dataset.
}
\description{
This dataset is a subset of Fishnet.
}
\usage{data("fishnet")}
\format{
  A data frame with 4455 observations and 2 variables.
  \describe{
    \item{\code{POINT_X}}{Longitude}
    \item{\code{POINT_Y}}{Latitude}
  }
}
\examples{
data(fishnet)
}
\keyword{datasets}
