% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsSamplingAbundance}
\alias{GibbsSamplingAbundance}
\title{Gibbs Sampling for LDA Abundance with Stick-Breaking}
\usage{
GibbsSamplingAbundance(DATA, nCommunity, beta, gamma, nGibbs,
  logLikelihoodAndPrior = TRUE, display_progress = TRUE)
}
\arguments{
\item{DATA}{- DataFrame with Abundance}

\item{beta}{- NumericVector for beta (Sx1)}

\item{gamma}{- Hyperparameter  Beta(1,gamma)}

\item{nGibbs}{- Total number of Gibbs Samples}

\item{logLikelihoodAndPrior}{- Likelihood compute with Priors ?}

\item{int}{nCommunity - Number of communities}

\item{bool}{display_progress=true - Should I Show the progressBar ?}
}
\value{
List - With Theta(nGibbs,nLocations*nCommunity), Phi(nGibbs,nCommunity*nSpecies) and logLikelihood
}
\description{
Compute the Gibbs Sampling for LDA Abundance with Stick-Breaking
}

