% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsSamplingPresence}
\alias{GibbsSamplingPresence}
\title{Gibbs Sampling for LDA Presence and Absence}
\usage{
GibbsSamplingPresence(DATA, nCommunity, alpha0, alpha1, gamma, nGibbs,
  logLikelihoodAndPrior = TRUE, display_progress = TRUE)
}
\arguments{
\item{DATA}{- DataFrame with Presence and Absecence (Zeros and Ones)}

\item{alpha0}{- Hyperparameter Beta(alpha0,alpha1)}

\item{alpha1}{- Hyperparameter Beta(alpha0,alpha1)}

\item{gamma}{- Hyperparameter  Beta(1,gamma)}

\item{nGibbs}{- Total number of Gibbs Samples}

\item{logLikelihoodAndPrior}{- Likelihood compute with Priors ?}

\item{int}{nCommunity - Number of communities}

\item{bool}{display_progress=true - Should I Show the progressBar ?}
}
\value{
List - With Theta(nGibbs,nCommunity*nSpecies), Phi(nGibbs,nLocations*nCommunity) and logLikelihood
}
\description{
Compute the Gibbs Sampling for LDA Presence and Absence
}

