\name{Landsat}
\alias{Landsat}
\docType{data}
\title{ Landsat TM 5 imagery from 2010 of the Iquitos-Nauta road in the Peruvian Amazon }
\description{
This data set has Binomial data from Landsat TM 5 imagery from 2010 of the Iquitos-Nauta road in the Peruvian Amazon for 7 bands at 69540 locations.
}
\usage{data(Landsat)}
\format{
  A data frame with 69540 observations for 9 columns.
}
\source{
This dataset is from:
Valle D, Baiser B,Woodall CW, Chazdon R (2014). "Decomposing biodiversity data using the
Latent Dirichlet Allocation model, a probabilistic multivariate statistical method." Ecology
letters, 17(12), 1591-1601.
}
\examples{
data(Landsat)
}
\keyword{datasets}
