\name{LocationsBirds}
\alias{LocationsBirds}
\docType{data}
\title{
ID variable Latitude and Longitude Locations for Birds dataset.
}
\description{
This dataset is a subset of Birds dataset with ID variable.
}
\usage{data("LocationsBirds")}
\format{
  A data frame with 3080 observations and 3 variables.
  \describe{
    \item{\code{loc.id}}{Location ID}
    \item{\code{Latitude}}{Latitude}
    \item{\code{Longitude}}{Longitude}
  }
}
\examples{
data(LocationsBirds)
}
\keyword{datasets}
