\name{presence}
\alias{presence}
\docType{data}
\title{ Species Presence/Absence Data }
\description{
This data set has Presence/Absence predictions for 13 species at 386 forested locations. It consists of species, observed presence-absence values, and the probability predictions of three different models.
}
\usage{data(presence)}
\format{
  A data frame with 386 observations for 13 species.
  Each cell represents one when the specie is presented zero otherwise.
}
\source{
This dataset is from:
Moisen, G.G., Freeman, E.A., Blackard, J.A., Frescino, T.S., Zimmerman N.E., Edwards, T.C. Predicting tree species presence and basal area in Utah: A comparison of stochastic gradient boosting, generalized additive models, and tree-based methods. Ecological Modellng, 199 (2006) 176-187.   
}
\examples{
data(presence)
}
\keyword{datasets}