\name{sp500}
\alias{sp500}
\docType{data}
\title{ Daily transactions Sp500 }
\description{
Daily transactions for 46 firms of the Sp500 Index in 2015.
}
\usage{data(sp500)}
\format{
  A data frame with 249 observations for 46 firms.
  Each cell represents one when some transactions ocurred and zero otherwise.
  }
\examples{
data(sp500)
}
\keyword{datasets}
