\name{complaints}
\alias{complaints}
\docType{data}
\title{
Complaints received for the **Bureau of Consumer Financial Protection** in US about financial products and services.
}
\description{
Specifically in this dataset we work with only credit card complaint's for the 2015 year.
}
\usage{data("complaints")}
\format{
  A data frame with 17301 observations on the following 3 variables.
  \describe{
    \item{\code{Product}}{a factor with levels \code{Credit card}}
    \item{\code{Issue}}{a factor with levels \code{Advertising and marketing} \code{Application processing delay} \code{APR or interest rate} \code{Arbitration} \code{Balance transfer} \code{Balance transfer fee} \code{Bankruptcy} \code{Billing disputes} \code{Billing statement} \code{Cash advance} \code{Cash advance fee} \code{Closing/Cancelling account} \code{Convenience checks} \code{Credit card protection / Debt protection} \code{Credit determination} \code{Credit line increase/decrease} \code{Customer service / Customer relations} \code{Delinquent account} \code{Forbearance / Workout plans} \code{Identity theft / Fraud / Embezzlement} \code{Late fee} \code{Other} \code{Other fee} \code{Overlimit fee} \code{Payoff process} \code{Privacy} \code{Rewards} \code{Sale of account} \code{Transaction issue} \code{Unsolicited issuance of credit card}}
    \item{\code{Company}}{a factor with levels \code{Adir International, LLC} \code{Affiliates Management Company} \code{Afni, Inc.} \code{AllianceOne Recievables Management} \code{Alliant Capital Management LLC} \code{Allied Interstate LLC} \code{Ally Financial Inc.} \code{Alpha Recovery Corp} \code{Altisource Portfolio Solutions, S.a r.l.} \code{American Coradius International LLC} \code{Amex} \code{Amscot Corporation} \code{Army and Air Force Exchange Service} \code{Arvest Bank} \code{Asset Recovery Solutions, LLC} \code{Atlanticus Services Corporation} \code{Avant Credit Corporation} \code{Banco Popular de Puerto Rico} \code{Banco Santander Puerto Rico} \code{BancorpSouth Bank} \code{Bank of America} \code{Bank of China USA} \code{Bank of the West} \code{Bankers Healthcare Group, LLC} \code{Barclays PLC} \code{BB&T Financial} \code{BBVA Compass} \code{Blackhawk Network Holdings Inc.} \code{Blatt, Hasenmiller, Leibsker & Moore, LLC} \code{BMO Harris} \code{BNY Mellon} \code{Boeing Employees' Credit Union} \code{BOK Financial Corp} \code{Brightwater Capital, LLC} \code{Bursey & Associates, P.C.} \code{Capital Management Services, LP} \code{Capital One} \code{Castle Credit Co Holdings} \code{Cavalry Investments, LLC} \code{Cedars Business Services, LLC} \code{Central Portfolio Control Inc.} \code{Cheek Law Offices, LLC} \code{Citibank} \code{Citizens Financial Group, Inc.} \code{CL Holdings, LLC} \code{Colony Brands, Inc.} \code{Comerica} \code{Commerce Bank} \code{Commercial Recovery Systems} \code{Conn's, Inc.} \code{Consumer Asset Management (Texas)} \code{Consumer Defense Group, LLC} \code{Continental Finance Company, LLC} \code{Convergent Resources, Inc.} \code{Convergys} \code{CoreLogic} \code{Couch Conville & Blitt LLC} \code{Credit Card Receivables Fund Incorporated} \code{Credit Corp Solutions Inc.} \code{Credit Technologies, Inc.} \code{D&A Services, LLC} \code{Daniels Norelli Scully & Cecere, P.C.} \code{Delbert Services} \code{Dennis Dillon Auto Park and Truck Center, Inc.,} \code{Discover} \code{Duvera Billing Services, LLC} \code{Dynamic Recovery Solutions, LLC} \code{Echelon Financial, Inc.} \code{Element Global Services} \code{Emmett L. Goodman, Jr., LLC} \code{Empowerment Ventures, LLC} \code{Encore Capital Group} \code{Equifax} \code{ERC} \code{Experian} \code{Expert Global Solutions, Inc.} \code{Fenton & McGarvey Law Firm, P.S.C.} \code{Fidelity National Information Services, Inc. (FNIS)} \code{Fifth Third Financial Corporation} \code{Finance Store Inc.} \code{Financial Credit Service, Inc.} \code{Finkelstein, Kern, Steinberg & Cunningham, PC} \code{First Citizens} \code{First Data Corporation} \code{First National Bank of Omaha} \code{First National Collection Bureau, Inc.} \code{First Niagara Bank} \code{First Step Group, LLC} \code{First Tennessee Bank} \code{FirstBank of Puerto Rico} \code{FirstMerit Bank} \code{Flagstar Bank} \code{Flurish Inc.} \code{FMA Alliance, Ltd.} \code{FMS Inc.} \code{Forster & Garbus LLP} \code{Frederick J. Hanna & Associates, P.C.} \code{FTC} \code{Gatestone & Co. International Inc.} \code{Genesis Lending} \code{Global Credit & Collection Corporation} \code{GreenSky Trade Credit, LLC} \code{Halsted Financial Services, LLC.} \code{Hameroff Law Group, P.C.} \code{HSBC North America Holdings Inc.} \code{IBERIABANK} \code{Investment Retrievers, Inc.} \code{Investors Bank} \code{JH Portfolio Debt Equities, LLC} \code{Jim Bottin Enterprises, Inc} \code{JPMorgan Chase & Co.} \code{KeyBank NA} \code{Kirkland Asset Management LLC} \code{Kirschenbaum & Phillips P.C} \code{Law Office of Michael J. Scott, PC} \code{Law Offices Howard Lee Schiff, P.C.} \code{Law Offices of Ed Overcash, LLC} \code{Legal Prevention Services, LLC.} \code{Legal Recovery Law Offices, Inc} \code{LHR Inc.} \code{Liberty Credit Services, Inc.} \code{LoanDepot} \code{Lobel Financial Corporation} \code{Logan View LLC} \code{Love, Beal & Nixon, P.C.} \code{LTD Financial Services, L.P.} \code{M&T Bank Corporation} \code{Machol & Johannes, LLC} \code{Mapother & Mapother, PSC} \code{Mark A. Kirkorsky, PC} \code{Marlette Funding, LLC} \code{Mel S. Harris and Associates, LLC} \code{Meyer & Njus, P.A.} \code{Mid-Atlantic Consumer Services} \code{MoneyGram} \code{Morgan Stanley} \code{MTACC Inc.} \code{Mullooly Jeffrey Rooney & Flynn} \code{National Enterprise Systems, Inc.} \code{Nations Recovery Center} \code{Nationwide Debt Management Solutions, LLC} \code{Navy FCU} \code{Nelson Cruz & Associates LLC} \code{NetSpend Corporation,  a TSYS Company} \code{New York Community Bank} \code{Northstar Location Services, LLC} \code{One Technologies, LP} \code{OneMain Financial Holdings, LLC} \code{Operating Partners Co., LLC} \code{Oracle Financial Group LLC.} \code{P&L Acquisitions LLC} \code{Patenaude & Felix APC} \code{PayPal Holdings, Inc.} \code{Penn Credit Corporation} \code{Pentagon FCU} \code{People's United Bank} \code{Phillips & Cohen Associates, Ltd.} \code{Pinnacle Recovery, Inc.} \code{Plastyc, Inc.} \code{Platinum Holdings Group, LLC} \code{Plaza Home Mortgage} \code{PNC Bank N.A.} \code{Portfolio Recovery Associates, Inc.} \code{Praxis Financial Solutions, Incorporated} \code{Preferred Credit, Inc.} \code{Pressler & Pressler, LLP} \code{Progressive Financial Services, Inc.} \code{Prosperity Bancshares, Inc} \code{Protas, Spivok & Collins, LLC} \code{Radius Global Solutions LLC} \code{Real Time Resolutions} \code{Receivables Performance Management LLC} \code{Regional Management Corp.} \code{Regions Financial Corporation} \code{Resurgent Capital Services L.P.} \code{Rossman & Co & PCB} \code{Sacor Financial, Inc} \code{Santander Bank US} \code{Santander Consumer USA Holdings Inc} \code{SchoolsFirst Federal Credit Union} \code{Scott Fetzer Financial Group, Inc.} \code{Second Round Limited Partnership} \code{Sentry Credit, Inc.} \code{SIMM Associates, Inc.} \code{Smith Debnam Narron Drake Saintsing & Myers LLP} \code{SquareTwo Financial Corporation} \code{State Employee's Credit Union} \code{State Farm Bank} \code{Stenger & Stenger, P.C} \code{Sterling Jewelers Inc.} \code{Stillman Law Office} \code{SunTrust Banks, Inc.} \code{Suttell, Hammer & White, P.S} \code{Synchrony Financial} \code{Synovus Bank} \code{Tate and Kirlin Assoc} \code{TCF National Bank} \code{TD Bank US Holding Company} \code{The Huntington National Bank} \code{The Law Firm of Allan C. Smith} \code{The Law Office of John P. Frye P.C.} \code{Thomas, King & Associates} \code{Tidewater Finance Company} \code{Total Card, Inc.} \code{Toyota Motor Credit Corporation} \code{TransUnion Intermediate Holdings, Inc.} \code{Transworld Systems Inc.} \code{Troy Capital, LLC} \code{Trustmark Corporation} \code{Turning Point Solutions LLC} \code{U.S. Bancorp} \code{UBS Bank} \code{UMB Bank} \code{Union Bank} \code{United Debt Holdings, LLC} \code{United Revenue Corporation} \code{URS Holding, LLC} \code{USAA Savings} \code{Vion Holdco I LLC} \code{WAKEFIELD & ASSOCIATES, INC.} \code{Wells Fargo & Company} \code{Weltman, Weinberg & Reis} \code{Western Alliance Bancorporation} \code{Whitney Bank} \code{Xoom} \code{Zions Bancorporation}}
  }
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
\url{http://catalog.data.gov/dataset/consumer-complaint-database}
}
\examples{
data(complaints)
}
\keyword{datasets}
